Fire tile, player status updated
Fire tiles now properly damage over time.
Instantaneous and timed healing is the same as damage, so it is also affected.
Note that HP is tracked as a decimal but is reported as an int. This allows damage or healing over time less than 1 to have an effect.
Reducing HP to less than 1 makes HP 0, raising it to over max HP makes it max HP.
When HP is zero, no other effects matter until the player's "status.alive" is set to true.